Make attribute setting an inner join (#45)
* Add support for .value in expressions

* Reformat with black

* Cleanup and use an inner join

* Reformat with black
